Splitted tests under modules
